+++ pkg/plugin/init.go
@@ -27,6 +27,7 @@ import (
 	"github.com/wklken/apisix-go/pkg/plugin/request_context"
 	"github.com/wklken/apisix-go/pkg/plugin/request_id"
 	"github.com/wklken/apisix-go/pkg/plugin/request_validation"
+	"github.com/wklken/apisix-go/pkg/plugin/syslog"
 	"github.com/wklken/apisix-go/pkg/plugin/ua_restriction"
 	"github.com/wklken/apisix-go/pkg/plugin/udp_logger"
 	"github.com/wklken/apisix-go/pkg/plugin/uri_blocker"
@@ -86,6 +87,8 @@ func New(name string) Plugin {
 		return &http_logger.Plugin{}
 	case "udp-logger":
 		return &udp_logger.Plugin{}
+	case "syslog":
+		return &syslog.Plugin{}
 	}
 	return nil
 }