+++ CHANGELOG.md
@@ -59,3 +59,4 @@
 - [x] plugin consumer-restriction [consumer-restriction](https://apisix.apache.org/zh/docs/apisix/plugins/consumer-restriction/) 2024-04-29
 - [x] plugin http-logger [http-logger](https://apisix.apache.org/zh/docs/apisix/plugins/http-logger/) 2024-04-30
 - [x] plugin udp-logger [udp-logger](https://apisix.apache.org/zh/docs/apisix/plugins/udp-logger/) 2024-05-01
+- [x] plugin sys-logger [syslog](https://apisix.apache.org/zh/docs/apisix/plugins/syslog/) 2024-05-01